+++ tse_analytics/modules/phenomaster/io/tse_dataset_loader.py
@@ -1,10 +1,8 @@
 import json
 import sqlite3
 import timeit
-from datetime import timedelta
 from pathlib import Path
 
-import numpy as np
 import pandas as pd
 from loguru import logger
 
@@ -13,19 +11,14 @@ from tse_analytics.core.data.datatable import Datatable
 from tse_analytics.core.data.shared import Aggregation, Animal, Variable
 from tse_analytics.modules.phenomaster.data.phenomaster_dataset import PhenoMasterDataset
 from tse_analytics.modules.phenomaster.data.predefined_variables import assign_predefined_values
-from tse_analytics.modules.phenomaster.io.tse_import_settings import TseImportSettings
-from tse_analytics.modules.phenomaster.submodules.actimot.data.actimot_data import ActimotData
-from tse_analytics.modules.phenomaster.submodules.calo.data.calo_data import CaloData
-from tse_analytics.modules.phenomaster.submodules.drinkfeed.data.drinkfeed_data import DrinkFeedData
-
-CHUNK_SIZE = 1000000
-
-ACTIMOT_RAW_TABLE = "actimot_raw"
-DRINKFEED_BIN_TABLE = "drinkfeed_bin"
-CALO_BIN_TABLE = "calo_bin"
+from tse_analytics.modules.phenomaster.io import tse_import_settings
+from tse_analytics.modules.phenomaster.submodules.actimot.io.data_loader import read_actimot_raw
+from tse_analytics.modules.phenomaster.submodules.calo.io.data_loader import read_calo_bin
+from tse_analytics.modules.phenomaster.submodules.drinkfeed.io.data_loader import read_drinkfeed_bin, read_drinkfeed_raw
+from tse_analytics.modules.phenomaster.submodules.grouphousing.io.data_loader import read_grouphousing
 
 
-def load_tse_dataset(path: Path, import_settings: TseImportSettings) -> PhenoMasterDataset | None:
+def load_tse_dataset(path: Path, import_settings: tse_import_settings.TseImportSettings) -> PhenoMasterDataset | None:
     """
     Loads and processes a PhenoMaster dataset from the specified path and settings.
 
@@ -73,21 +66,32 @@ def load_tse_dataset(path: Path, import_settings: TseImportSettings) -> PhenoMas
 
     # Import ActoMot raw data if present
     if import_settings.import_actimot_raw:
-        if ACTIMOT_RAW_TABLE in metadata["tables"]:
-            actimot_data = _read_actimot_raw(path, dataset)
+        if tse_import_settings.ACTIMOT_RAW_TABLE in metadata["tables"]:
+            actimot_data = read_actimot_raw(path, dataset)
             dataset.actimot_data = actimot_data
 
     # Import drinkfeed bin data if present
     if import_settings.import_drinkfeed_bin:
-        if DRINKFEED_BIN_TABLE in metadata["tables"]:
-            drinkfeed_data = _read_drinkfeed_bin(path, dataset)
-            dataset.drinkfeed_data = drinkfeed_data
+        if tse_import_settings.DRINKFEED_BIN_TABLE in metadata["tables"]:
+            drinkfeed_bin_data = read_drinkfeed_bin(path, dataset)
+            dataset.drinkfeed_bin_data = drinkfeed_bin_data
+
+    if import_settings.import_drinkfeed_raw:
+        if tse_import_settings.DRINKFEED_RAW_TABLE in metadata["tables"]:
+            drinkfeed_raw_data = read_drinkfeed_raw(path, dataset)
+            dataset.drinkfeed_raw_data = drinkfeed_raw_data
 
     # Import calo bin data if present
     if import_settings.import_calo_bin:
-        if CALO_BIN_TABLE in metadata["tables"]:
-            calo_data = _read_calo_bin(path, dataset)
+        if tse_import_settings.CALO_BIN_TABLE in metadata["tables"]:
+            calo_data = read_calo_bin(path, dataset)
             dataset.calo_data = calo_data
+
+    # Import group housing data if present
+    if import_settings.import_grouphousing:
+        if tse_import_settings.GROUP_HOUSING_TABLE in metadata["tables"]:
+            grouphousing_data = read_grouphousing(path, dataset)
+            dataset.grouphousing_data = grouphousing_data
 
     logger.info(f"Import complete in {(timeit.default_timer() - tic):.3f} sec: {path}")
 
@@ -118,14 +122,14 @@ def _read_metadata(path: Path) -> dict:
     metadata["experiment"]["runtime"] = str(pd.to_timedelta(metadata["experiment"]["runtime"], unit="ms"))
     metadata["experiment"]["cycle_interval"] = str(pd.to_timedelta(metadata["experiment"]["cycle_interval"], unit="ms"))
 
-    if "main_table" in metadata["tables"]:
-        metadata["tables"]["main_table"]["sample_interval"] = str(
-            pd.to_timedelta(metadata["tables"]["main_table"]["sample_interval"], unit="ms")
+    if tse_import_settings.MAIN_TABLE in metadata["tables"]:
+        metadata["tables"][tse_import_settings.MAIN_TABLE]["sample_interval"] = str(
+            pd.to_timedelta(metadata["tables"][tse_import_settings.MAIN_TABLE]["sample_interval"], unit="ms")
         )
 
-    if "actimot_raw" in metadata["tables"]:
-        metadata["tables"]["actimot_raw"]["sample_interval"] = str(
-            pd.to_timedelta(metadata["tables"]["actimot_raw"]["sample_interval"], unit="ms")
+    if tse_import_settings.ACTIMOT_RAW_TABLE in metadata["tables"]:
+        metadata["tables"][tse_import_settings.ACTIMOT_RAW_TABLE]["sample_interval"] = str(
+            pd.to_timedelta(metadata["tables"][tse_import_settings.ACTIMOT_RAW_TABLE]["sample_interval"], unit="ms")
         )
 
     return metadata
@@ -141,6 +145,11 @@ def _get_animals(data: dict) -> dict[str, Animal]:
             "Text2": item["text2"],
             "Text3": item["text3"],
         }
+
+        # Add Tag property if exists
+        if "Tag" in item:
+            properties["Tag"] = item["Tag"]
+
         animal = Animal(
             enabled=bool(item["enabled"]),
             id=str(item["id"]),
@@ -155,7 +164,7 @@ def _read_main_table(
     path: Path,
     dataset: PhenoMasterDataset,
 ) -> tuple[pd.DataFrame, dict[str, Variable], pd.Timedelta]:
-    metadata = dataset.metadata["tables"]["main_table"]
+    metadata = dataset.metadata["tables"][tse_import_settings.MAIN_TABLE]
 
     sample_interval = pd.Timedelta(metadata["sample_interval"])
 
@@ -178,10 +187,10 @@ def _read_main_table(
     df = pd.DataFrame()
     with sqlite3.connect(path, check_same_thread=False) as connection:
         for chunk in pd.read_sql_query(
-            "SELECT * FROM main_table",
+            f"SELECT * FROM {tse_import_settings.MAIN_TABLE}",
             connection,
             dtype=dtypes,
-            chunksize=CHUNK_SIZE,
+            chunksize=tse_import_settings.CHUNK_SIZE,
         ):
             df = pd.concat([df, chunk], ignore_index=True)
 
@@ -209,221 +218,3 @@ def _read_main_table(
     variables = dict(sorted(variables.items(), key=lambda x: x[0].lower()))
 
     return df, variables, sample_interval
-
-
-def _read_actimot_raw(path: Path, dataset: PhenoMasterDataset) -> ActimotData:
-    metadata = dataset.metadata["tables"][ACTIMOT_RAW_TABLE]
-
-    sample_interval = pd.Timedelta(metadata["sample_interval"])
-
-    # Read variables list
-    skipped_variables = ["DateTime", "Box", "X1", "X2", "Y1", "Y2", "Z1", "Z2"]
-    variables: dict[str, Variable] = {}
-    dtypes = {}
-    for item in metadata["columns"].values():
-        variable = Variable(
-            item["id"],
-            item["unit"],
-            item["description"],
-            item["type"],
-            Aggregation.MEAN,
-            False,
-        )
-        if variable.name not in skipped_variables:
-            variables[variable.name] = variable
-        dtypes[variable.name] = item["type"]
-    # Ignore the time for "DateTime" column
-    dtypes.pop("DateTime")
-
-    # Read measurements data
-    df = pd.DataFrame()
-    with sqlite3.connect(path, check_same_thread=False) as connection:
-        for chunk in pd.read_sql_query(
-            f"SELECT * FROM {ACTIMOT_RAW_TABLE}",
-            connection,
-            dtype=dtypes,
-            chunksize=CHUNK_SIZE,
-        ):
-            chunk["X"] = np.left_shift(chunk["X2"].to_numpy(dtype=np.uint64), 32) + chunk["X1"].to_numpy(
-                dtype=np.uint64
-            )
-            chunk.rename(columns={"Y1": "Y"}, inplace=True)
-            chunk.drop(columns=["X1", "X2"], inplace=True)
-            df = pd.concat([df, chunk], ignore_index=True)
-
-    # Convert DateTime from POSIX format
-    df["DateTime"] = pd.to_datetime(df["DateTime"], origin="unix", unit="ns")
-
-    actimot_data = ActimotData(
-        dataset,
-        ACTIMOT_RAW_TABLE,
-        str(path),
-        variables,
-        df,
-        sample_interval,
-    )
-
-    return actimot_data
-
-
-def _read_drinkfeed_bin(path: Path, dataset: PhenoMasterDataset) -> DrinkFeedData:
-    metadata = dataset.metadata["tables"][DRINKFEED_BIN_TABLE]
-
-    # Read variables list
-    skipped_variables = ["DateTime", "Box"]
-    variables: dict[str, Variable] = {}
-    dtypes = {}
-    for item in metadata["columns"].values():
-        variable = Variable(
-            item["id"],
-            item["unit"],
-            item["description"],
-            item["type"],
-            Aggregation.MEAN,
-            False,
-        )
-        if variable.name not in skipped_variables:
-            variables[variable.name] = variable
-        dtypes[variable.name] = item["type"]
-    # Ignore the time for "DateTime" column
-    dtypes.pop("DateTime")
-
-    # Read measurements data
-    df = pd.DataFrame()
-    with sqlite3.connect(path, check_same_thread=False) as connection:
-        for chunk in pd.read_sql_query(
-            f"SELECT * FROM {DRINKFEED_BIN_TABLE}",
-            connection,
-            dtype=dtypes,
-            chunksize=CHUNK_SIZE,
-        ):
-            df = pd.concat([df, chunk], ignore_index=True)
-
-    # Convert DateTime from POSIX format
-    df["DateTime"] = pd.to_datetime(df["DateTime"], origin="unix", unit="ns")
-
-    # Add Animal column
-    box_to_animal_map = {}
-    for animal in dataset.animals.values():
-        box_to_animal_map[animal.properties["Box"]] = animal.id
-    df["Animal"] = df["Box"].replace(box_to_animal_map)
-    df = df.astype({
-        "Animal": "category",
-    })
-
-    drinkfeed_data = DrinkFeedData(
-        dataset,
-        DRINKFEED_BIN_TABLE,
-        str(path),
-        variables,
-        df,
-    )
-
-    return drinkfeed_data
-
-
-def _read_calo_bin(path: Path, dataset: PhenoMasterDataset) -> CaloData:
-    metadata = dataset.metadata["tables"][CALO_BIN_TABLE]
-
-    sample_interval = pd.Timedelta(metadata["sample_interval"])
-
-    # Read variables list
-    skipped_variables = ["DateTime", "Box"]
-    variables: dict[str, Variable] = {}
-    dtypes = {}
-    for item in metadata["columns"].values():
-        variable = Variable(
-            item["id"],
-            item["unit"],
-            item["description"],
-            item["type"],
-            Aggregation.MEAN,
-            False,
-        )
-        if variable.name not in skipped_variables:
-            variables[variable.name] = variable
-        dtypes[variable.name] = item["type"]
-    # Ignore the time for "DateTime" column
-    dtypes.pop("DateTime")
-
-    # Read measurements data
-    df = pd.DataFrame()
-    with sqlite3.connect(path, check_same_thread=False) as connection:
-        for chunk in pd.read_sql_query(
-            f"SELECT * FROM {CALO_BIN_TABLE}",
-            connection,
-            dtype=dtypes,
-            chunksize=CHUNK_SIZE,
-        ):
-            df = pd.concat([df, chunk], ignore_index=True)
-
-    # Convert DateTime from POSIX format
-    df["DateTime"] = pd.to_datetime(df["DateTime"], origin="unix", unit="ns")
-
-    # Sort dataframe
-    df.sort_values(["Box", "DateTime"], inplace=True)
-
-    # Assign bins
-    previous_timestamp = None
-    previous_box = None
-    bins = []
-    offsets = []
-    timedeltas = []
-    time_gap = timedelta(seconds=10)
-    offset = 0
-    for row in df.itertuples():
-        timestamp = row.DateTime
-        box = row.Box
-
-        if box != previous_box:
-            start_timestamp = timestamp
-
-        if previous_timestamp is None:
-            bins = [0]
-        elif timestamp - previous_timestamp > time_gap:
-            bin_number = bins[-1]
-            bin_number = bin_number + 1
-
-            offset = 0
-
-            # reset bin number for a new box
-            if box != previous_box:
-                bin_number = 0
-
-            bins.append(bin_number)
-            start_timestamp = timestamp
-        else:
-            bin_number = bins[-1]
-
-            # reset bin number for a new box
-            if box != previous_box:
-                bin_number = 0
-                offset = 0
-
-            bins.append(bin_number)
-
-        if box != previous_box:
-            previous_box = box
-
-        td = timestamp - start_timestamp
-        timedeltas.append(td)
-
-        # offset = td.total_seconds()
-        offsets.append(offset)
-        offset = offset + 1
-        previous_timestamp = timestamp
-
-    df.insert(1, "Timedelta", timedeltas)
-    df.insert(2, "Bin", bins)
-    df.insert(3, "Offset", offsets)
-
-    calo_data = CaloData(
-        dataset,
-        CALO_BIN_TABLE,
-        str(path),
-        variables,
-        df,
-        sample_interval,
-    )
-
-    return calo_data