+++ tse_analytics/modules/phenomaster/submodules/actimot/io/data_loader.py
@@ -1,12 +1,69 @@
+import sqlite3
 from pathlib import Path
 
 import numpy as np
 import pandas as pd
 
 from tse_analytics.core.csv_import_settings import CsvImportSettings
-from tse_analytics.core.data.shared import Variable
+from tse_analytics.core.data.shared import Variable, Aggregation
+from tse_analytics.modules.phenomaster.io import tse_import_settings
 from tse_analytics.modules.phenomaster.submodules.actimot.data.actimot_data import ActimotData
 from tse_analytics.modules.phenomaster.data.phenomaster_dataset import PhenoMasterDataset
+
+
+def read_actimot_raw(path: Path, dataset: PhenoMasterDataset) -> ActimotData:
+    metadata = dataset.metadata["tables"][tse_import_settings.ACTIMOT_RAW_TABLE]
+
+    sample_interval = pd.Timedelta(metadata["sample_interval"])
+
+    # Read variables list
+    skipped_variables = ["DateTime", "Box", "X1", "X2", "Y1", "Y2", "Z1", "Z2"]
+    variables: dict[str, Variable] = {}
+    dtypes = {}
+    for item in metadata["columns"].values():
+        variable = Variable(
+            item["id"],
+            item["unit"],
+            item["description"],
+            item["type"],
+            Aggregation.MEAN,
+            False,
+        )
+        if variable.name not in skipped_variables:
+            variables[variable.name] = variable
+        dtypes[variable.name] = item["type"]
+    # Ignore the time for "DateTime" column
+    dtypes.pop("DateTime")
+
+    # Read measurements data
+    df = pd.DataFrame()
+    with sqlite3.connect(path, check_same_thread=False) as connection:
+        for chunk in pd.read_sql_query(
+            f"SELECT * FROM {tse_import_settings.ACTIMOT_RAW_TABLE}",
+            connection,
+            dtype=dtypes,
+            chunksize=tse_import_settings.CHUNK_SIZE,
+        ):
+            chunk["X"] = np.left_shift(chunk["X2"].to_numpy(dtype=np.uint64), 32) + chunk["X1"].to_numpy(
+                dtype=np.uint64
+            )
+            chunk.rename(columns={"Y1": "Y"}, inplace=True)
+            chunk.drop(columns=["X1", "X2"], inplace=True)
+            df = pd.concat([df, chunk], ignore_index=True)
+
+    # Convert DateTime from POSIX format
+    df["DateTime"] = pd.to_datetime(df["DateTime"], origin="unix", unit="ns")
+
+    actimot_data = ActimotData(
+        dataset,
+        tse_import_settings.ACTIMOT_RAW_TABLE,
+        str(path),
+        variables,
+        df,
+        sample_interval,
+    )
+
+    return actimot_data
 
 
 def import_actimot_csv_data(
@@ -18,7 +75,7 @@ def import_actimot_csv_data(
     return None
 
 
-def _load_from_csv(path: Path, dataset: PhenoMasterDataset, csv_import_settings: CsvImportSettings):
+def _load_from_csv(path: Path, dataset: PhenoMasterDataset, csv_import_settings: CsvImportSettings) -> ActimotData:
     with open(path) as f:
         lines = f.readlines()