+++ tse_analytics/modules/phenomaster/submodules/drinkfeed/data/drinkfeed_raw_data.py
@@ -1,0 +1,23 @@
+import pandas as pd
+
+from tse_analytics.core.data.shared import Variable
+
+
+class DrinkFeedRawData:
+    def __init__(
+        self,
+        dataset,
+        name: str,
+        path: str,
+        variables: dict[str, Variable],
+        df: pd.DataFrame,
+    ):
+        self.dataset = dataset
+        self.name = name
+        self.path = path
+        self.variables = variables
+        self.raw_df = df
+
+    @property
+    def start_timestamp(self):
+        return self.raw_df.at[0, "DateTime"]