+++ tse_analytics/modules/phenomaster/submodules/grouphousing/trafficage_config.py
@@ -1,0 +1,8 @@
+TRAFFICAGE_POSITIONS = {
+    "TraffiCage1": (0, 0),
+    "TraffiCage2": (0, 1),
+    "TraffiCage3": (1, 0),
+    "TraffiCage4": (1, 1),
+    "TraffiCage5": (2, 0),
+    "TraffiCage6": (2, 1),
+}