+++ tse_analytics/modules/phenomaster/views/import_tse_dialog_ui.py
@@ -3,7 +3,7 @@
 ################################################################################
 ## Form generated from reading UI file 'import_tse_dialog.ui'
 ##
-## Created by: Qt User Interface Compiler version 6.9.0
+## Created by: Qt User Interface Compiler version 6.9.2
 ##
 ## WARNING! All changes made in this file will be lost when recompiling UI file!
 ################################################################################
@@ -41,10 +41,20 @@ class Ui_ImportTseDialog(object):
 
         self.verticalLayout_3.addWidget(self.checkBoxDrinkFeedBin)
 
+        self.checkBoxDrinkFeedRaw = QCheckBox(self.groupBoxImportSettings)
+        self.checkBoxDrinkFeedRaw.setObjectName(u"checkBoxDrinkFeedRaw")
+
+        self.verticalLayout_3.addWidget(self.checkBoxDrinkFeedRaw)
+
         self.checkBoxActiMotRaw = QCheckBox(self.groupBoxImportSettings)
         self.checkBoxActiMotRaw.setObjectName(u"checkBoxActiMotRaw")
 
         self.verticalLayout_3.addWidget(self.checkBoxActiMotRaw)
+
+        self.checkBoxGroupHousing = QCheckBox(self.groupBoxImportSettings)
+        self.checkBoxGroupHousing.setObjectName(u"checkBoxGroupHousing")
+
+        self.verticalLayout_3.addWidget(self.checkBoxGroupHousing)
 
 
         self.horizontalLayout.addWidget(self.groupBoxImportSettings)
@@ -72,6 +82,8 @@ class Ui_ImportTseDialog(object):
         self.groupBoxImportSettings.setTitle(QCoreApplication.translate("ImportTseDialog", u"Import", None))
         self.checkBoxCaloBin.setText(QCoreApplication.translate("ImportTseDialog", u"Calo bin data", None))
         self.checkBoxDrinkFeedBin.setText(QCoreApplication.translate("ImportTseDialog", u"DrinkFeed bin data", None))
+        self.checkBoxDrinkFeedRaw.setText(QCoreApplication.translate("ImportTseDialog", u"DrinkFeed raw data", None))
         self.checkBoxActiMotRaw.setText(QCoreApplication.translate("ImportTseDialog", u"ActiMot raw data", None))
+        self.checkBoxGroupHousing.setText(QCoreApplication.translate("ImportTseDialog", u"Group Housing data", None))
     # retranslateUi